move all of templateservicebroker under a single package
