+++ pkg/templateservicebroker/servicebroker/catalog.go
@@ -10,8 +10,8 @@ import (
 	jsschema "github.com/lestrrat/go-jsschema"
 
 	oapi "github.com/openshift/origin/pkg/api"
-	"github.com/openshift/origin/pkg/openservicebroker/api"
 	templateapi "github.com/openshift/origin/pkg/template/apis/template"
+	"github.com/openshift/origin/pkg/templateservicebroker/openservicebroker/api"
 )
 
 const (
