+++ pkg/templateservicebroker/servicebroker/provision.go
@@ -4,6 +4,8 @@ import (
 	"net/http"
 	"reflect"
 
+	"github.com/golang/glog"
+
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
@@ -11,10 +13,9 @@ import (
 	kapi "k8s.io/kubernetes/pkg/api"
 	"k8s.io/kubernetes/pkg/apis/authorization"
 
-	"github.com/golang/glog"
 	"github.com/openshift/origin/pkg/authorization/util"
-	"github.com/openshift/origin/pkg/openservicebroker/api"
 	templateapi "github.com/openshift/origin/pkg/template/apis/template"
+	"github.com/openshift/origin/pkg/templateservicebroker/openservicebroker/api"
 	uservalidation "github.com/openshift/origin/pkg/user/apis/user/validation"
 )
