+++ pkg/templateservicebroker/servicebroker/test-scripts/bind.sh
