+++ pkg/templateservicebroker/servicebroker/test-scripts/catalog.sh
