+++ pkg/templateservicebroker/servicebroker/test-scripts/clusterrole.yaml
