+++ pkg/templateservicebroker/servicebroker/test-scripts/provision.sh
