+++ pkg/templateservicebroker/servicebroker/unbind.go
@@ -3,15 +3,16 @@ package servicebroker
 import (
 	"net/http"
 
+	"github.com/golang/glog"
+
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apiserver/pkg/authentication/user"
 	"k8s.io/kubernetes/pkg/apis/authorization"
 
-	"github.com/golang/glog"
 	"github.com/openshift/origin/pkg/authorization/util"
-	"github.com/openshift/origin/pkg/openservicebroker/api"
 	templateapi "github.com/openshift/origin/pkg/template/apis/template"
+	"github.com/openshift/origin/pkg/templateservicebroker/openservicebroker/api"
 )
 
 // Unbind is the reverse of Bind.  Currently it simply removes the binding ID
